Create README.md
Receivers: The gloves made by Nike are engineered to offer the highest level of dexterity and ball control. To safeguard the hands and fingers, they have carefully placed padding and lightweight, breathable materials.Enter your desired changes in this section. Then, use the button below to paraphrase. It really is that simple!

Running backs: The gloves made by Nike are made to be strong and protective. To survive the demands of the running game, they have reinforced stitching and thicker padding.

Linemen: Power and grip are the two main features of Nike’s lineman’s gloves. Their rough palms and robust padding aid linemen in maintaining control over the line of scrimmage.

The Nike Youth Football Gloves’ features